Update skeleton components folder path
